requirements.txt
Django==4.2.5
django-environ==0.11.2
django-allauth==0.61.0
django-debug-toolbar==4.2.0
psycopg==3.1.12
django-extensions==3.2.3
gunicorn==23.0.0
whitenoise==6.7.0
pillow==10.4.0
django-cloudinary-storage==0.3.0
django-compressor==4.4
django-sass-processor==1.4
libsass==0.22.0
flake8==7.0.0
